Move logic out of Definition class
